INCLUDE INFERENCE
In the search of valid frame groups, if a frame is valid as a First
Frame (H>0), the rest of the frames infer the keypoints that have
confidence < threshold.

Modified logic in dataprocessor to take into account the logic
difference between been suitable as a first frame and as a
non first frame.